backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
openai==1.6.1
gunicorn==21.2.0